\name{maizeA}
\docType{data}
\alias{maizeA}
\title{Numerator relationship matrix required for "gwerAM()"}
\description{
  This matrix contains the pairwise relatedness information required for \command{gwerAM}.
}

\usage{data(maizeA)}
\format{The matrix has 277 rows and columns. Each cell gives the relatedness
of the corresponding pair of maize inbred lines. This information has been derived from \code{data(maizeW)} of 277 diverse maize inbreds
which have been used in the study of Yu et al. (2006).
}

%\source{}
\references{
Yu, J., Pressoir, G., Briggs, W.H., Bi, I.V., Yamasaki, M., Doebley, J.F., McMullen, M.D., Gaut, B.S., Holland, J.B., Kresovich, S., and Buckler, E.S. 2006. A unified mixed-model method for association mapping accounting for multiple levels of relatedness. Nature Genetics 38: 203-208.
  }


