\name{maizeQ}
\docType{data}
\alias{maizeQ}
\title{Population structure matrix required for "gwerAM()"}
\description{
  This matrix contains the population structure information required for \command{gwerAM}.
}

\usage{data(maizeQ)}
\format{The matrix has five columns, one for each subpopulation. Each row gives the probability that the corresponding
genotype belongs to the corresponding subpopulations. The data set comprises information from 277 diverse maize inbreds
which have been used in the study of Yu et al. (2006).
}

%\source{}
\references{
Yu, J., Pressoir, G., Briggs, W.H., Bi, I.V., Yamasaki, M., Doebley, J.F., McMullen, M.D., Gaut, B.S., Holland, J.B., Kresovich, S., and Buckler, E.S. 2006. A unified mixed-model method for association mapping accounting for multiple levels of relatedness. Nature Genetics 38: 203-208.
  }


