\name{maizevarianceA}
\docType{data}
\alias{maizevarianceA}
\title{Genotypic variance required for "gwerAM()"}
\description{
  This value is the genotypic variance of the trait days to pollen (Yu et al. 2006) of a diverse set of 277 maize inbred lines as inferred from QK mixed model analysis without marker data. This information is required for \command{gwerAM}.
}

\usage{data(maizevarianceA)}
\format{This data set comprises only one value.
}

%\source{}
\references{
Yu, J., Pressoir, G., Briggs, W.H., Bi, I.V., Yamasaki, M., Doebley, J.F., McMullen, M.D., Gaut, B.S., Holland, J.B., Kresovich, S., and Buckler, E.S. 2006. A unified mixed-model method for association mapping accounting for multiple levels of relatedness. Nature Genetics 38: 203-208.
  }


